Orphan Alley, Tags, Doorway Script
Got started on building up Orphan Alley, placed a few placeholder sprites there. Cemetery Door is set up but not working currently. Tried to figure out a useful way to order tags, failed, killed a few as a result. Started a new Doorway script as an alternate method to teleport the player somewhere.